Move config module
