+++ server/src/app.module.ts
@@ -3,8 +3,8 @@ import { TypeOrmModule, TypeOrmModuleOptions } from '@nestjs/typeorm';
 import { GraphQLModule } from '@nestjs/graphql';
 import { AppController } from './app.controller';
 import { AppService } from './app.service';
-import { ConfigModule } from './packages-old/config/config.module';
-import { ConfigService } from './packages-old/config/config.service';
+import { ConfigModule } from './modules/config/config.module';
+import { ConfigService } from './modules/config/config.service';
 import { AuthModule } from './packages-old/auth/auth.module';
 import { SubmissionModule } from './modules/reviewer/submission.module';
 import { PassportModule } from '@nestjs/passport';
