+++ server/src/main.ts
@@ -1,7 +1,7 @@
 import { NestFactory } from '@nestjs/core';
 import { NestExpressApplication } from '@nestjs/platform-express';
 import { AppModule } from './app.module';
-import { ConfigService } from './packages-old/config/config.service';
+import { ConfigService } from './modules/config/config.service';
 import { join } from 'path';
 
 async function bootstrap() {
